fix(expr): PostgreSQL semantics for three-valued logic, integer division, SUBSTR and casts
Three ADR-0012 conformance fixes found by the PostgreSQL differential
oracle. All seven knownBug pins are deleted and the entries gated again:
the oracle now holds 94 gated / 7 pinned, from 87 / 14.

BoolNullExpr protocol (EvalBoolNull → val, null) carries UNKNOWN through
Cmp (all variants), the connectives, IN/NOT IN (list, uncorrelated and
correlated subquery — with a NULL-in-set flag), BETWEEN, LIKE, and the
shape-specialized nodes. The boxed Eval path boxes UNKNOWN as SQL NULL;
EvalBool remains the WHERE collapse (keeps neither FALSE nor UNKNOWN).
And/Or keep their two-valued short-circuit EvalBool — collapsing before
or after min/max logic agrees — so the filter hot path pays nothing. NOT
is the operator where the orders disagree, and it now evaluates on the
three-valued protocol: `1 NOT IN (2, NULL)` no longer admits a row SQL
excludes. IS [NOT] TRUE/FALSE moves off Cmp onto a dedicated IsBool
NULL-test node, since NULL IS TRUE is false, not UNKNOWN.

former float-`/` pin followed DuckDB, which ADR-0012 overturns. The rule
lands in all three arithmetic layers: BinOpInt64 (literals), BinOpNumeric
(mode-resolved columns), and the generic BinOp (untyped operands — CASTs,
negated columns). Truncation is semantics, not optimization, so it does
not ride the int-arith kill switch: with WADJET_INT_ARITH=0 the float
delegate's quotient is truncated (divTrunc, structural operand check).
Unary minus preserves integer identity (fold of negated literals at
compile; int64 negation at eval; numeric declared type in the planner,
which also makes ORDER BY -col sort numerically instead of by rendered
text). intArithAllInt mirrors the runtime by declaring Int64 for `/`.
The SUM(x/k)→SUM(x)/k rewrite declines for integer literal divisors,
where per-row truncation breaks linearity. Division by zero stays NULL
(no error channel; the missing 22012 remains pinned on the wire arm).

Blast radius, measured: no TPC-H answer moves (the suite's only `/` has
a float literal; SF0.01 and SF1 fingerprints hold, SF1 62s ≈ baseline).
The DuckDB stored baseline is byte-identical except the new entry. Two
engine answers change, both now matching PostgreSQL: SUM(v/2) over an
int column truncates per row (wadjet suite), and ip/2 returns int64
(test suite). The DuckDB arm handles the dialect divergence per ADR-0012
rule 3 — a duckdbSQL override asks DuckDB the same question in its `//`
spelling (IntegerDivisionDialect), never exempting the entry.

below 1 consumes length before the string begins: SUBSTR('abcdef', 0, 3)
is 'ab'. One substrWindow clamp shared by the scalar and vectorized
kernels (the vec kernel also no longer slices out of bounds on negative
lengths). CAST of a fractional value to an integer type rounds half away
from zero instead of truncating — TRUNC() is the truncating spelling —
in expr.Cast and the cast_int scalar.

Gates: go vet clean; gofmt clean on changed files; internal suite 42/42;
wadjet, test, full benchmarks/tpch (DuckDB stored + live compare,
two-path invariance, optimization invariance) green; task pg-oracle:test
green with the seven pins deleted.

Fixes #369
Fixes #370
Fixes #373

+++ test/polymorphic_column_typing_test.go
@@ -129,13 +129,17 @@ func TestPolymorphicFunctionsOverUnresolvableColumns(t *testing.T) {
 			want: []any{"ALGERIA", "ARGENTINA", "BRAZIL"},
 		},
 		{
-			// No scan carries "max(n_nationkey)", so the fallback stands and
-			// the output is the declared Float64 — the residual this change
-			// deliberately does not reach, pinned so it cannot become a wrong
-			// CONFIDENT answer without a test noticing.
+			// No scan carries "max(n_nationkey)", so the aggregate argument
+			// cannot decide — the literal beside it does. `-1` decides Int64
+			// exactly as `1` always has, now that unary ± carries its
+			// operand's numeric type through nodeDeclaredType (#369):
+			// PostgreSQL types this int, and before that rule the sign alone
+			// demoted it to the Float64 fallback. The aggregate-output
+			// residual itself is unchanged — an aggregate arg still answers
+			// "undecided", which is what this entry pins.
 			name: "an aggregate output is not a scan column",
 			sql:  `SELECT COALESCE(MAX(n_nationkey), -1) AS x FROM nation`,
-			want: []any{float64(2)},
+			want: []any{int64(2)},
 		},
 	}
 	for _, tc := range tests {
